Replace base/matrixStats/DelayedArray by DelayedMatrixStat (#66)
This is a straight find and replace of (col|row)(Sums|Means) with DelayedMatrixStats equivalents. Immediately, this is to work around an apparent bug in DelayedArray,rowSums-method (https://github.com/Bioconductor/DelayedArray/issues/16) but long term want to be using the optimised implementations in DelayedMatrixStat (e.g., using `cols` and `rows` args).