Small tweak to "updateObject" method for BSseq objects
The method now calls the "updateObject" method for RangedSummarizedExperiment
objects (via callNextMethod()) early rather than late. The earlier the better
because very few things actually work in general on an outdated BSseq object
before that call is made. In particular the endoapply() loop that now follows
the call to callNextMethod() was failing if the BSseq object had "old"
DelayedArray objects in its assays.