Use a single IPC lock when writing M and Cov to RealizationSink
Previously, process A could write to 'M' and process B could write to 'Cov' simultaneously, but this can create a race condition if, for example, both 'M' and 'Cov' are in the same HDF5 file.
This change means that, in this example, process 'A' must wait for process 'B' to finish before being able to begin writing 'M' to the RealizationSink.